backend/requirements.txt
flask
flask-cors
pandas
python-pptx
openpyxl
gunicorn
psycopg2-binary
